requirements.txt
psutil>=5.9.5
python-dotenv==1.0.0
torch>=2.8.0
transformers>=4.33.2
requests>=2.31.0
tweepy>=4.14.0
